+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/codestyle/EmptyControlStatementRule.java
@@ -7,18 +7,26 @@ package net.sourceforge.pmd.lang.java.rule.codestyle;
 import net.sourceforge.pmd.lang.java.ast.ASTBlock;
 import net.sourceforge.pmd.lang.java.ast.ASTDoStatement;
 import net.sourceforge.pmd.lang.java.ast.ASTEmptyStatement;
+import net.sourceforge.pmd.lang.java.ast.ASTExpression;
+import net.sourceforge.pmd.lang.java.ast.ASTFieldAccess;
 import net.sourceforge.pmd.lang.java.ast.ASTFinallyClause;
 import net.sourceforge.pmd.lang.java.ast.ASTForStatement;
 import net.sourceforge.pmd.lang.java.ast.ASTForeachStatement;
 import net.sourceforge.pmd.lang.java.ast.ASTIfStatement;
 import net.sourceforge.pmd.lang.java.ast.ASTInitializer;
+import net.sourceforge.pmd.lang.java.ast.ASTLocalVariableDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTResource;
 import net.sourceforge.pmd.lang.java.ast.ASTResourceList;
+import net.sourceforge.pmd.lang.java.ast.ASTSuperExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTSwitchStatement;
 import net.sourceforge.pmd.lang.java.ast.ASTSynchronizedStatement;
+import net.sourceforge.pmd.lang.java.ast.ASTThisExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTTryStatement;
+import net.sourceforge.pmd.lang.java.ast.ASTVariableAccess;
+import net.sourceforge.pmd.lang.java.ast.ASTVariableId;
 import net.sourceforge.pmd.lang.java.ast.ASTWhileStatement;
 import net.sourceforge.pmd.lang.java.ast.JavaNode;
+import net.sourceforge.pmd.lang.java.ast.internal.PrettyPrintingUtil;
 import net.sourceforge.pmd.lang.java.rule.AbstractJavaRulechainRule;
 import net.sourceforge.pmd.lang.java.rule.internal.JavaRuleUtil;
 import net.sourceforge.pmd.properties.PropertyDescriptor;
@@ -133,27 +141,52 @@ public class EmptyControlStatementRule extends AbstractJavaRulechainRule {
     public Object visit(ASTTryStatement node, Object data) {
         if (isEmpty(node.getBody())) {
             // all resources must be explicitly ignored
-            boolean allResourcesIgnored = true;
             boolean hasResource = false;
             ASTResourceList resources = node.getResources();
             if (resources != null) {
                 for (ASTResource resource : resources) {
                     hasResource = true;
-                    String name = resource.getStableName();
+                    ASTLocalVariableDeclaration localVarDecl = resource.asLocalVariableDeclaration();
+                    if (localVarDecl == null) {
+                        // not a concise resource.
+                        ASTExpression init = resource.getInitializer();
+                        if (isSimpleExpression(init)) {
+                            // The expression is simple enough, it should be just written this.close() or var.close(),
+                            // so we report this case
+                            asCtx(data).addViolationWithMessage(node, "Empty try-with-resources statement. Should be written {0}.close()", PrettyPrintingUtil.prettyPrint(init));
+                            return null;
+                        }
+                        // Otherwise the expression is more complex and this is allowed, in order
+                        // to avoid bringing a variable into the enclosing scope
+                        continue;
+                    }
+
+                    // A named resource. Named should be ignored.
+                    ASTVariableId varId = localVarDecl.getVarIds().firstOrThrow();
+                    if (!varId.getLocalUsages().isEmpty()) {
+                        // this resource is used in other resource declarations or in a finally block
+                        return null;
+                    }
+                    String name = varId.getName();
                     if (!JavaRuleUtil.isExplicitUnusedVarName(name)) {
-                        allResourcesIgnored = false;
-                        break;
+                        asCtx(data).addViolationWithMessage(node, "Empty try-with-resources statement. Rename the resource to `ignored`, `unused` or `_` (Java 22+).");
+                        return null;
                     }
                 }
             }
 
-            if (hasResource && !allResourcesIgnored) {
-                asCtx(data).addViolationWithMessage(node, "Empty try body - you could rename the resource to ''ignored''");
-            } else if (!hasResource) {
+            if (!hasResource) {
                 asCtx(data).addViolationWithMessage(node, "Empty try body");
             }
         }
         return null;
+    }
+
+    private static boolean isSimpleExpression(ASTExpression init) {
+        return init instanceof ASTThisExpression
+            || init instanceof ASTSuperExpression
+            || init instanceof ASTVariableAccess
+            || init instanceof ASTFieldAccess && isSimpleExpression(((ASTFieldAccess) init).getQualifier());
     }
 
     private boolean isEmpty(JavaNode node) {